Remove system download manager dependency from DownloadInfoRepository
